Handled review comments
