+++ integration/spark2/src/main/scala/org/apache/spark/util/AlterTableUtil.scala
@@ -29,6 +29,7 @@ import org.apache.carbondata.common.logging.LogService
 import org.apache.carbondata.core.constants.CarbonCommonConstants
 import org.apache.carbondata.core.locks.{CarbonLockFactory, ICarbonLock, LockUsage}
 import org.apache.carbondata.core.metadata.schema.table.CarbonTable
+import org.apache.carbondata.core.util.CarbonProperties
 import org.apache.carbondata.format.{SchemaEvolutionEntry, TableInfo}
 
 object AlterTableUtil {
@@ -88,6 +89,36 @@ object AlterTableUtil {
     locks.foreach { carbonLock =>
       if (carbonLock.unlock()) {
         LOGGER.info("Alter table lock released successfully")
+      } else {
+        LOGGER.error("Unable to release lock during alter table operation")
+      }
+    }
+  }
+
+  /**
+   * This method will release the locks by manually forming a lock path. Specific usage for
+   * rename table
+   *
+   * @param locks
+   * @param locksAcquired
+   * @param dbName
+   * @param tableName
+   * @param storeLocation
+   * @param LOGGER
+   */
+  def releaseLocksManually(locks: List[ICarbonLock],
+      locksAcquired: List[String],
+      dbName: String,
+      tableName: String,
+      storeLocation: String,
+      LOGGER: LogService): Unit = {
+    val lockLocation = storeLocation + CarbonCommonConstants.FILE_SEPARATOR +
+                       dbName + CarbonCommonConstants.FILE_SEPARATOR + tableName
+    locks.zip(locksAcquired).foreach { case (carbonLock, lockType) =>
+      val lockFilePath = lockLocation + CarbonCommonConstants.FILE_SEPARATOR +
+                         lockType
+      if (carbonLock.releaseLockManually(lockFilePath)) {
+        LOGGER.info(s"Alter table lock released successfully: ${ lockType }")
       } else {
         LOGGER.error("Unable to release lock during alter table operation")
       }
