+++ integration/spark2/src/main/scala/org/apache/spark/util/Compaction.scala
@@ -20,7 +20,7 @@ import org.apache.spark.sql.{CarbonEnv, SparkSession}
 import org.apache.spark.sql.execution.command.{AlterTableCompaction, AlterTableModel}
 
 import org.apache.carbondata.core.constants.CarbonCommonConstants
-import org.apache.carbondata.spark.merger.CompactionType
+import org.apache.carbondata.processing.merger.CompactionType
 
 /**
  * table compaction api
