+++ processing/src/main/java/org/apache/carbondata/processing/merger/NodeBlockRelation.java
@@ -14,7 +14,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.apache.carbondata.spark.merger;
+package org.apache.carbondata.processing.merger;
 
 import org.apache.carbondata.core.datastore.block.Distributable;
