+++ processing/src/main/java/org/apache/carbondata/processing/merger/exeception/SliceMergerException.java
@@ -15,7 +15,7 @@
  * limitations under the License.
  */
 
-package org.apache.carbondata.spark.merger.exeception;
+package org.apache.carbondata.processing.merger.exeception;
 
 import java.util.Locale;
